refactor: init + commands modules + add luadoc
